requirements.txt
#core
Django==4.1.3
djangorestframework==3.14.0
setuptools==65.6.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
gunicorn==20.1.0

# image, pdf
pdfminer.six==20220524
Pillow==9.2.0

django-ninja==0.19.1
